fix(ml): close the genai.Client after using it
It looks like we're not the only one having networking issues with
google-genai: https://github.com/pydantic/pydantic-ai/issues/3071.

The lib now explicitely advices to close the connection:
https://github.com/googleapis/python-genai?tab=readme-ov-file#close-a-client

That's what this PR is about:

- close the underlying HTTP connection using a with block
- don't reuse the client, create a new one each time
- update the Credentials object directly, instead of letting Google
  creating a new one every-time. This is similar to what is done in
  https://ai.pydantic.dev/models/google/#service-account

For reference, we need to provide the `project` to Google genai.Client
now, as otherwise it ignores the credentials we provided and tries
to find credential files at the default location.

I also upgraded google-genai, as new versions fixed some of the
networking issues that were brought up in the issue tracker.

At first, I started to integrate pydantic-ai, but as pydantic-ai relies
directly on google-genai for making the requests, let's stick to
google-genai for now.

This is working locally, let's see of it improves "Network unreachable"
issues.

toml linter also failed on pyproject.toml, I run it to solve the issue.